+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `shape(n)` method
 - Changed `NArray` to inherit from `Rice::Object`
+- Changed `read_ptr` and `write_ptr` methods to raise exception for non-contiguous arrays
 
 ## 0.3.1 (2026-02-12)